Modifications to salt generation process:
This commit switches the generation of `S` to be more inline with the
design.  Instead of concating the base64 encoded `s_0` with the javascript
string of the username, we instead convert `s_0` to raw bytes, then encode
the javascript string of the username into utf-8 bytes, then concat the two.